Migrate updated publish form
