+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@heroicons/react": "^2.0.18",
     "@metamask/sdk-react": "^0.27.0",
+    "dayjs": "^1.11.13",
     "flowbite-react": "^0.7.2",
     "formik": "^2.4.6",
     "next": "^14.1.4",
@@ -19,10 +20,12 @@
     "react-dom": "^18.2.0",
     "react-jwt": "^1.2.0",
     "react-tag-input-component": "^2.0.2",
+    "react-tailwindcss-datepicker": "^1.7.2",
     "sharp": "^0.33.5",
     "yup": "^1.4.0"
   },
   "devDependencies": {
+    "@tailwindcss/forms": "^0.5.9",
     "@types/node": "20.4.6",
     "@types/react": "18.2.18",
     "autoprefixer": "10.4.12",
